+++ UnityProject/Assets/Resources/Prefabs/Items/Cargo.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 99aaeed6cfb76e54db0a1609ca1313f4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
