+++ UnityProject/Assets/Scripts/Cargo.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9f80e80ef508fb543b4ff22cbdad2809
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
