--- UnityProject/Assets/Scripts/Objects/PosterBehaviour.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 815148f04bd24613abae9285eea1e68d
-timeCreated: 1572196618
