+++ UnityProject/Assets/Scripts/Objects/Posters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2ca39cac511cdb241a6746dea1414cb7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
