+++ UnityProject/Assets/Scripts/Objects/Posters/Poster.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2ac0c24eef5d19046bdab404f15a958e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
