+++ UnityProject/Assets/Scripts/Objects/Posters/PosterBehaviour.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 815148f04bd24613abae9285eea1e68d
+timeCreated: 1572196618
